+++ cxx/Protocol.cpp
@@ -51,7 +51,7 @@ namespace Zeze
 				ByteBuffer pbb(os.Bytes, os.ReadIndex, size);
 				os.ReadIndex += size;
 				Service::ProtocolFactoryHandle factoryHandle;
-				long long type = ((long long)moduleId << 32) | (unsigned int)protocolId;
+				int64_t type = ((int64_t)moduleId << 32) | (unsigned int)protocolId;
 				if (service->FindProtocolFactoryHandle(type, factoryHandle))
 				{
 					std::auto_ptr<Protocol> p(factoryHandle.Factory());
